v3.0.1.0 - renamed getFormattedNmsDouble -> getFormattedNmsInteger (had no reason to be a double, plus floating point might've been an issue in the future) - Added KamiCommon commands, and a test suite for NMS

+++ spigot-nms/src/main/java/com/kamikazejam/kamicommon/nms/provider/BlockUtilProvider.java
@@ -10,7 +10,7 @@ import org.jetbrains.annotations.NotNull;
  */
 public class BlockUtilProvider extends Provider<AbstractBlockUtil> {
     @Override
-    protected @NotNull AbstractBlockUtil provide(double ver) {
+    protected @NotNull AbstractBlockUtil provide(int ver) {
         if (ver < f("1.8")) {
             throw new IllegalArgumentException("Version not supported (< 1.8): " + ver);
         }
